add Chart

+++ app/Http/Controllers/APIGheChieuController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Models\GheChieu;
 use App\Models\PhongChieu;
+use App\Models\QuyenChucNang;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 
@@ -13,6 +15,18 @@ class APIGheChieuController extends Controller
 {
     public function store(Request $request)
     {
+        $id_chuc_nang   =   25;
+        $user_login     =   Auth::guard('admin')->user();
+
+        $check          =   QuyenChucNang::where('id_quyen', $user_login->id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if (!$check) {
+            return response()->json([
+                'status'    => 0,
+                'message'   => 'Bạn không có quyền cho chức năng này!',
+            ]);
+        }
         DB::beginTransaction();
         try {
             GheChieu::where('id_phong_chieu', $request->id)->delete();
@@ -45,6 +59,19 @@ class APIGheChieuController extends Controller
     }
     public function info(Request $request)
     {
+        $id_chuc_nang   =   26;
+        $user_login     =   Auth::guard('admin')->user();
+
+        $check          =   QuyenChucNang::where('id_quyen', $user_login->id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if (!$check) {
+            return response()->json([
+                'status'    => 0,
+                'message'   => 'Bạn không có quyền cho chức năng này!',
+            ]);
+        }
+
         DB::beginTransaction();
         try {
 
@@ -63,6 +90,18 @@ class APIGheChieuController extends Controller
     }
     public function status(Request $request)
     {
+        $id_chuc_nang   =   27;
+        $user_login     =   Auth::guard('admin')->user();
+
+        $check          =   QuyenChucNang::where('id_quyen', $user_login->id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if (!$check) {
+            return response()->json([
+                'status'    => 0,
+                'message'   => 'Bạn không có quyền cho chức năng này!',
+            ]);
+        }
         DB::beginTransaction();
 
         try {
@@ -91,6 +130,18 @@ class APIGheChieuController extends Controller
 
     public function update(Request $request)
     {
+        $id_chuc_nang   =   28;
+        $user_login     =   Auth::guard('admin')->user();
+
+        $check          =   QuyenChucNang::where('id_quyen', $user_login->id_quyen)
+            ->where('id_chuc_nang', $id_chuc_nang)
+            ->first();
+        if (!$check) {
+            return response()->json([
+                'status'    => 0,
+                'message'   => 'Bạn không có quyền cho chức năng này!',
+            ]);
+        }
         DB::beginTransaction();
         try {
